[UPDATE] Home컨트롤러 API 분리
